% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scala.R
\name{make_sbt_engine}
\alias{make_sbt_engine}
\title{An sbt engine which can be used to create scala engine in knitr.}
\usage{
make_sbt_engine(path)
}
\arguments{
\item{path}{the directory in which the sbt probject will be created.}
}
\description{
An sbt engine which can be used to create scala engine in knitr.
}
