% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scala.R
\name{make_scala_engine}
\alias{make_scala_engine}
\title{Scala engine for knitr.}
\usage{
make_scala_engine(...)
}
\arguments{
\item{...}{params passed to \code{\link{scala}} function from rscala package.}
}
\description{
Scala engine for knitr.
}
